Fix: use fork of `gatsby-source-magento2`
The official community source plugin for Magento2 has a bug that causes errors when used with Gatsby 4.

I created a fork, fixed the bug and filed a PR with the official package - until that PR is resolved, we need to use my fork of the source plugin for it to work with Gatsby 4.